+++ hata/env/__init__.py
@@ -58,6 +58,7 @@ HATA_STATUS_ENDPOINT : `None`, `str` = `None`
     Discord status endpoint.
 """
 
+from .getter_context import *
 from .getters import *
 from .loading import *
 from .parsing import *
@@ -65,6 +66,7 @@ from .variables import *
 
 
 __all__ = (
+    *getter_context.__all__,
     *getters.__all__,
     *loading.__all__,
     *parsing.__all__,